updates to start-over.sh script
Attempt to fix issues with the 'verifyDiffIsOnlyMods' function.
Messy for now since I couldn't figure out why it was failing at first.

The issue turned out to be that grep returns an error status code when
no lines are found (returns '1'). This caused a failure of the script
because I have 'set -e' activated (i.e., exit immediately if a pipeline,
list, or compound command, exits with a non-zero status). But an empty
git diff is exactly the state I am hoping for. So now I am temporarily
disabling "set -e" (with "set +e")

I also had a typo on the 'git diff' line, a 'k' after the semicolon.
I'm think I didn't notice/check that as an issue before modifying. eek