Attempt a different technique to pull in source commit files
This technique is also not correct. I don't know why I can't get my mind
wrapped propery around this. Maybe a 'git reset --soft' is the correct
thing to do.

Also, need to handle merge commits...right now I pull the full git
history in as though it's purely linear. That will be a more complicated
thing to do, to be sure :D